feat: App 내 뽑기 버튼 기능 구현
